Split off LiveData data type hierarchy into separate module
